+++ jpostman-annotations/src/main/java/io/jpostman/annotations/runtime/JUnitPostmanFramework.java
@@ -12,41 +12,49 @@ import io.jpostman.junit.JUnitContext;
  */
 public final class JUnitPostmanFramework implements JPostmanFramework<JUnitContext> {
 
+	/** {@inheritDoc} */
 	@Override
 	public Class<JUnitContext> contextType() {
 		return JUnitContext.class;
 	}
 
+	/** {@inheritDoc} */
 	@Override
 	public JUnitContext create() {
 		return JUnitContext.create();
 	}
 
+	/** {@inheritDoc} */
 	@Override
 	public void setCurrent(JUnitContext context) {
 		JUnitContext.setCurrent(context);
 	}
 
+	/** {@inheritDoc} */
 	@Override
 	public void clearCurrent() {
 		JUnitContext.clearCurrent();
 	}
 
+	/** {@inheritDoc} */
 	@Override
 	public void secret(JUnitContext context, Object environment) {
 		context.secret((Environment) environment);
 	}
 
+	/** {@inheritDoc} */
 	@Override
 	public void load(JUnitContext context, InputStream rules) throws Exception {
 		context.load(rules);
 	}
 
+	/** {@inheritDoc} */
 	@Override
 	public JUnitContext loadRules(JUnitContext context, String rule) {
 		return context.loadRules(rule);
 	}
 
+	/** {@inheritDoc} */
 	@Override
 	public JUnitContext filter(JUnitContext context, String... paths) {
 		if (paths == null || paths.length == 0) {
@@ -55,16 +63,19 @@ public final class JUnitPostmanFramework implements JPostmanFramework<JUnitConte
 		return context.filter(paths);
 	}
 
+	/** {@inheritDoc} */
 	@Override
 	public JUnitContext request(JUnitContext context, Request request) {
 		return context.request(request);
 	}
 
+	/** {@inheritDoc} */
 	@Override
 	public JUnitContext response(JUnitContext context, ApiExecutor executor) {
 		return context.response(executor);
 	}
 
+	/** {@inheritDoc} */
 	@Override
 	public void verify(JUnitContext context, int statusCode, boolean soft, boolean log) {
 		if (soft) {
@@ -75,16 +86,19 @@ public final class JUnitPostmanFramework implements JPostmanFramework<JUnitConte
 		context.asserts(log).verify(statusCode);
 	}
 
+	/** {@inheritDoc} */
 	@Override
 	public Object cache(JUnitContext context, String key) {
 		return context.cache(key);
 	}
 
+	/** {@inheritDoc} */
 	@Override
 	public void cache(JUnitContext context, String key, Object value) {
 		context.cache(key, value);
 	}
 
+	/** {@inheritDoc} */
 	@Override
 	public String name() {
 		return "JUnit";